# testdata/openapi.json
{
  "components": {
...
    "schemas": {
      "Endpoint": {
        "description": "An inbound webhook endpoint",
        "example": {
          "enabled": true,
          "id": "019c0123-4567-7890-abcd-ef1234567890",
          "inbound_url": "https://runlater.eu/in/ep_abc123def456ghi789jkl012mno345",
          "inserted_at": "2026-01-29T10:00:00Z",
          "name": "Stripe webhooks",
          "retry_attempts": 5,
          "slug": "ep_abc123def456ghi789jkl012mno345",
          "updated_at": "2026-01-29T10:00:00Z",
          "use_lane": true,
          "forward_urls": [
            "https://myapp.com/webhooks/stripe"
          ]
        },
        "properties": {
          "enabled": {
            "default": true,
            "description": "Whether the endpoint is active",
            "type": "boolean"
          },
          "id": {
            "description": "Endpoint ID",
            "format": "uuid",
            "type": "string"
          },
          "inbound_url": {
            "description": "Full inbound URL",
            "format": "uri",
            "type": "string"
          },
          "inserted_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "description": "Endpoint name",
            "type": "string"
          },
          "notify_on_failure": {
            "description": "Override org-level failure notification setting for forwarded events. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "notify_on_recovery": {
            "description": "Override org-level recovery notification setting for forwarded events. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "retry_attempts": {
            "default": 5,
            "description": "Number of retry attempts for forwarding (0-10)",
            "type": "integer"
          },
          "slug": {
            "description": "Unique slug for inbound URL",
            "type": "string"
          },
          "updated_at": {
            "description": "Last update timestamp",
            "format": "date-time",
            "type": "string"
          },
          "use_lane": {
            "default": true,
            "description": "Whether to use a named lane for serial execution",
            "type": "boolean"
          },
          "forward_urls": {
            "description": "URLs to forward events to (1-10)",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "forward_urls"
        ],
        "title": "Endpoint",
        "type": "object"
      },
      "EndpointRequest": {
        "description": "Request body for creating or updating an endpoint",
        "example": {
          "forward_url": "https://myapp.com/webhooks/stripe",
          "name": "Stripe webhooks",
          "retry_attempts": 5,
          "use_lane": true
        },
        "properties": {
          "enabled": {
            "default": true,
            "type": "boolean"
          },
          "forward_url": {
            "description": "URL to forward events to",
            "format": "uri",
            "type": "string"
          },
          "name": {
            "description": "Endpoint name",
            "type": "string"
          },
          "notify_on_failure": {
            "description": "Override org-level failure notification setting for forwarded events. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "notify_on_recovery": {
            "description": "Override org-level recovery notification setting for forwarded events. null = use org default.",
            "nullable": true,
            "type": "boolean"
          },
          "retry_attempts": {
            "default": 5,
            "description": "Number of retry attempts for forwarding (0-10)",
            "type": "integer"
          },
          "use_lane": {
            "default": true,
            "description": "Whether to use a named lane for serial execution",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "forward_url"
        ],
        "title": "EndpointRequest",
        "type": "object"
      },
      "EndpointResponse": {
        "description": "Response containing an endpoint",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Endpoint"
          }
        },
        "title": "EndpointResponse",
        "type": "object"
      },
      "EndpointsResponse": {
        "description": "Response containing a list of endpoints",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "type": "array"
          }
        },
        "title": "EndpointsResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Error response",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ExecutionsResponse": {
        "description": "Response containing execution history",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Execution"
            },
            "type": "array"
          }
        },
        "title": "ExecutionsResponse",
        "type": "object"
      },
      "InboundEvent": {
        "description": "An inbound webhook event received by an endpoint",
        "properties": {
          "id": {
            "description": "Event ID",
            "format": "uuid",
            "type": "string"
          },
          "method": {
            "description": "HTTP method of the incoming request",
            "type": "string"
          },
          "received_at": {
            "description": "When the event was received",
            "format": "date-time",
            "type": "string"
          },
          "source_ip": {
            "description": "IP address of the sender",
            "type": "string"
          },
          "task_ids": {
            "description": "Task IDs created for forwarding",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "status": {
            "description": "Aggregated forwarding status",
            "nullable": true,
            "type": "string"
          }
        },
        "title": "InboundEvent",
        "type": "object"
      },
      "InboundEventsResponse": {
        "description": "Response containing a list of inbound events",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/InboundEvent"
            },
            "type": "array"
          }
        },
        "title": "InboundEventsResponse",
        "type": "object"
      },
...
      "ReplayResponse": {
        "description": "Response from replaying an inbound event",
        "properties": {
          "data": {
            "properties": {
              "execution_id": {
                "format": "uuid",
                "type": "string"
              },
              "scheduled_for": {
                "format": "date-time",
                "type": "string"
              },
              "status": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "title": "ReplayResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "API key in format: pk_live_xxx.sk_live_yyy",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for managing scheduled tasks and webhooks.\n\n## Authentication\n\nAll API requests require a Bearer token in the Authorization header:\n\n```\nAuthorization: Bearer pk_live_xxx.sk_live_yyy\n```\n\nAPI keys can be created in the Runlater dashboard under Organization Settings.\n",
    "title": "Runlater API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/endpoints": {
      "get": {
        "callbacks": {},
        "description": "Returns all inbound webhook endpoints for the authenticated organization",
        "operationId": "PrikkeWeb.Api.EndpointController.index",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointsResponse"
                }
              }
            },
            "description": "Endpoints list"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all endpoints",
        "tags": [
          "Endpoints"
        ]
      },
      "post": {
        "callbacks": {},
        "description": "Creates a new inbound webhook endpoint",
        "operationId": "PrikkeWeb.Api.EndpointController.create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRequest"
              }
            }
          },
          "description": "Endpoint parameters",
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            },
            "description": "Endpoint created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an endpoint",
        "tags": [
          "Endpoints"
        ]
      }
    },
    "/api/v1/endpoints/{endpoint_id}/events": {
      "get": {
        "callbacks": {},
        "description": "Returns the inbound event history for an endpoint",
        "operationId": "PrikkeWeb.Api.EndpointController.events",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (1-100, default 50)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboundEventsResponse"
                }
              }
            },
            "description": "Events"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List inbound events",
        "tags": [
          "Endpoints"
        ]
      }
    },
    "/api/v1/endpoints/{endpoint_id}/events/{event_id}/replay": {
      "post": {
        "callbacks": {},
        "description": "Creates a new forwarding execution for an existing inbound event",
        "operationId": "PrikkeWeb.Api.EndpointController.replay",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "endpoint_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Event ID",
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Event replayed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Replay an inbound event",
        "tags": [
          "Endpoints"
        ]
      }
    },
    "/api/v1/endpoints/{id}": {
      "delete": {
        "callbacks": {},
        "description": "Deletes an endpoint and all its inbound events",
        "operationId": "PrikkeWeb.Api.EndpointController.delete",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Endpoint deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete an endpoint",
        "tags": [
          "Endpoints"
        ]
      },
      "get": {
        "callbacks": {},
        "description": "Returns a single inbound webhook endpoint by ID",
        "operationId": "PrikkeWeb.Api.EndpointController.show",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            },
            "description": "Endpoint"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get an endpoint",
        "tags": [
          "Endpoints"
        ]
      },
      "patch": {
        "callbacks": {},
        "description": "Updates an existing inbound webhook endpoint",
        "operationId": "PrikkeWeb.Api.EndpointController.update (2)",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRequest"
              }
            }
          },
          "description": "Endpoint parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            },
            "description": "Endpoint updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an endpoint",
        "tags": [
          "Endpoints"
        ]
      },
      "put": {
        "callbacks": {},
        "description": "Updates an existing inbound webhook endpoint",
        "operationId": "PrikkeWeb.Api.EndpointController.update",
        "parameters": [
          {
            "description": "Endpoint ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointRequest"
              }
            }
          },
          "description": "Endpoint parameters",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointResponse"
                }
              }
            },
            "description": "Endpoint updated"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an endpoint",
        "tags": [
          "Endpoints"
        ]
      }
    },
...
    "/api/v1/tasks/{task_id}/executions": {
      "get": {
        "callbacks": {},
        "description": "Returns the execution history for a task",
        "operationId": "PrikkeWeb.Api.TaskController.executions",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (1-100, default 50)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionsResponse"
                }
              }
            },
            "description": "Executions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get execution history",
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://runlater.eu",
      "variables": {}
    }
  ],
...
}